Added jsdoc-style comments, and prettified
