+++ fetcher.js
@@ -1,12 +1,32 @@
 "use strict";
 
+/* A simple HTTP request library with small footprint */
 const fetch = require("node-fetch");
 
+/* Class representing a wrapper around the fetch library */
 class Fetcher {
+  /**
+   * Initializes the wrapper class
+   */
   constructor() {
     this.fetch = fetch;
   }
-  get(url) {
+  /**
+   * Performs a GET HTTP method on the given URL and returns the html data
+   * @param  {string}  url The web endpoint whose html data we want to retrieve
+   * @return {Promise}     The promise containing the html data
+   */
+  async get(url) {
+    /**
+     * Setting up the options object for node-fetch
+     *
+     * The reason for setting the specific headers is to make the endpoint
+     * think that the request is coming from a browser
+     *
+     * The headers and their values are simply taken Mozilla developer console
+     *
+     * @type {Object}
+     */
     let options = {
       headers: {
         "User-Agent":
@@ -21,12 +41,15 @@ class Fetcher {
         "Cache-Control": "no-cache",
       },
     };
-    return this.fetch(url, options)
-      .catch((error) => {
-        error.type = "fetcher";
-        return Promise.reject(error);
-      })
-      .then((response) => response.text());
+
+    try {
+      /* Perform the operation and return the html data as string */
+      let response = await this.fetch(url, options);
+      return response.text();
+    } catch (error) {
+      error.type = "fetcher"; // Setting type in case we want to track the source of the error later on
+      throw error;
+    }
   }
 }
